refacto(validator): split validator package into multiple files
